Update GN copy file rule.
This merges the syntax of the script and copy file rules and adds lots of documentation.

In doing this I realized that the template handling and escaping was bad in the script rules, so I did a bunch of work and added much better tests.

BUG=288991
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/23536068

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b90c75afa26473fae4e6c2333beb241b7da27203