Remove ContainsOnlyWhitespace from string_util and CollapseWhitespace into to the base namespace
This function is only used in a few places and can trivially be implemented by passing the whitespace constant to the existing ContainsOnlyChars function.

This changes the ContainsOnlyChars function signature to take a StringPiece to avoid a copy from a literal to a standard string in the above-mentioned use-base.

Re-implement ContainsOnlyChars to use the find_first_not_of function on StringPiece.

BUG=
R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/183683024

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1e1229a168974388ff8cda5e0e066b5978ec227f