Add "." accessors to GN.
Add support for values referring to a Scope, and add support for the dot operator to read such values out of a scope.

This is not currently used for anything. A followup patch will use this for the template implementation.

R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/207233003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 51d0172bd506a8602337386e1c098f7b84238470