GN: Bug fixes in process_file_template
process_file_template would fail when the template argument was a string instead of a list, but the previous behavior (and the documentation) describes it as being able to take a string or a list as an argument.

It would also fail if the result was not an absolute file. This is required for most uses of sources expansions, but not for process_file_template which just returns the strings of whatever the results were. This adds "AsStrings" version of the sources expansions writer.

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/462103004

Cr-Original-Commit-Position: refs/heads/master@{#289112}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 209d521c49a8cae194ea5f1458a2ecbbba2bcc78