gn format: fix comments at end of blocks
Fixes comments at the end of a block like:

  stuff() {
    sources = []
    # wee
  }

This is done by storing an EndNode (instead of just a Token) as the
end of Block and List nodes. This means there's a node in the tree
traversal for the trailing ], and gives a place to attach the comment.

Also collapses some of the expression output code to reuse Sequence.

R=brettw@chromium.org
BUG=348474

Review URL: https://codereview.chromium.org/607173002

Cr-Original-Commit-Position: refs/heads/master@{#297226}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f7e13e52cddeda696e71c0ab4c08876ddeef4330