Fixes to {} handling in GN.
This disallows standalone {} blocks. Fortunately, we did not use these in
the Chrome build.

This eliminates the ambiguity of whether foo("bar") {} is a function that takes
a block (like a target definition) or a standalone function followed by a
standalone block.

Previously we would put the block on the function if a function preceeded it,
or keep the block standalone if there wasn't a function. Additionally,
functions that weren't expecting a block would silently ignore it, which
produced some weird cases.

This additionally adds checks for functions that don't expect blocks.

BUG=434519
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/1007963003

Cr-Original-Commit-Position: refs/heads/master@{#320749}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 989adc8450ef29ee2c902724323e68aa50e466b0