Add ninja dependency reference to gen written files.
For example, if one or more of the <blah>.tmp files is removed, even as part of an entire hierarchy/subtree currently the build fails because ninja does not know how to re-make it.

While it works, I don't like this hack.  It asserts that the build needs to be rewritten if files generated by "gn gen" are out of date.

While true, that's not really the best expression in the ninja 'language' of what's really happening.

It is a nice small change though.

BUG=469621
TEST=Build time only. See https://code.google.com/p/chromium/issues/detail?id=469621#c3

Review URL: https://codereview.chromium.org/1222693002

Cr-Original-Commit-Position: refs/heads/master@{#337873}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d82b2317e6bd619236ca7de934b341e7b32e7380