tools/gn: move the implementation of IsIdentifierXXX functions to the source file
Does not seem to have a reason to have some functions implemented in the
header while the others are in the source file.

This also allow us to remove the include of string_util.h, and have it
included only in the source file.

BUG=None
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/1483903004

Cr-Original-Commit-Position: refs/heads/master@{#362285}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0c60e3a9bb46ef9f3cf92f90f3e6fde598f3b8ee