Fix GN issues identified by MSVC static analysis.
MSVC's /analyze identified two missing null checks because it noted that a
pointer was null checked in one place but not another. After some study I
determined that null checks are not necessary for these values so I removed
them.

Although isxdigit takes an int, the input must be an unsigned char or the
behavior is undefined. I added the cast. I think this API is suboptimal.

BUG=643842,427616

Review-Url: https://codereview.chromium.org/2384493003
Cr-Original-Commit-Position: refs/heads/master@{#421945}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e23e7021f9a06d621088bc13856db9ba93c328c0