Setting "SubSystem" option on "Link" page for Visual Studio project.
When the debugger session in Visual Studio finished, the console window closed
by default. For console applications, such as unit tests, it is possible not to
close the console window immediately. This is controlled by the property
"SubSystem" on "Link" page in project settings. This patch set property
"SubSystem" in generated *.vcxproj file based on the "ldflags" field.

R=brettw@chromium.org
BUG=

Review-Url: https://chromiumcodereview.appspot.com/2428013002
Cr-Original-Commit-Position: refs/heads/master@{#426977}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2f8e5b9b27201976adb6f5526dad67b386872fbd