Move use_brlapi build flag to chrome.
It used to be in the global build faetures file but is only used by Chrome.

The define is removed. It was only used in one place to make sure the compile
matched the build flags, which in this case is uselessly circular.

Review-Url: https://codereview.chromium.org/2497653002
Cr-Original-Commit-Position: refs/heads/master@{#431665}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 519b7b6003aab8cd70808bb07f5d289e10149367