GN: Always check for bad items.
Previous checking for most types of errors in the GN introspection commands was
disabled. The reason was that it can be convenient to ask for certain types
of information which GN can provide even if there are errors elsewhere in the
build.

Unfortunately, not checking for bad items can leave the build graph in an
inconsistent state. In the case of missing dependencies, a target will never
be resolved which means most of the information the tools will return about it
will be missing or incorrect which is very surprising and confusing.

As a result, it seems best to always ensure things are in a consistent state
before running any introspection calculations.

Review-Url: https://codereview.chromium.org/2492273002
Cr-Original-Commit-Position: refs/heads/master@{#431667}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f77e13a490ea98bed6f136661fa778ba3972b8e1