Move if, else, true, false into a keyword category
This moves the if and else keywords out of the list of target
types and also marks true and false as keywords, bringing this
mode in line with the GN grammar.

Also remove a seemingly unnecessary line.

Change-Id: I9b28f7ca5bc60adae6302c9702960059e0285f1d
Reviewed-on: https://chromium-review.googlesource.com/540777
Reviewed-by: Elliot Glaysher <erg@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Kevin Peng <kevpeng@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#480900}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 18697c83b41fc147f21b373a55da451f0fd44117