GN: Make the linker output file depend on the inputs.
This is useful for files that are only read by the linker, such as
orderfiles. We were normally getting this right in the case where
the target contained at least one source file, as the output file
would transitively depend on the inputs via the source files, but if
the target only contained dependencies on other targets, the inputs
wouldn't be depended on at all.

Bug: 835622
Change-Id: Ic7fb39caeacd4698b7ec6cfb0cb583a429a117dd
Reviewed-on: https://chromium-review.googlesource.com/1041274
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Peter Collingbourne <pcc@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#555931}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ae62ba99a1a71db0758616215a557053d5fcf72a