format: Ensure that formatting always reaches a fixedpoint in one run
The .golden for this is ugly, but follows existing sources formatting e.g. in
https://gn.googlesource.com/gn/+/master/tools/gn/format_test_data/062.golden#25
This at least causes formatting to reach a fixed point in one step, which is
the more annoying bug.

In practice, the user will probably fix the sorting to the actual desired
formatting by manually sorting as was done in
https://chromium-review.googlesource.com/c/chromium/src/+/1403177.

Bug: gn:40

Test: CQ
Change-Id: Id3fea2564ae34156ac77e7445acd3f18bb4d161f
Reviewed-on: https://gn-review.googlesource.com/c/3720
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Scott Graham <scottmg@chromium.org>
Commit-Queue: Scott Graham <scottmg@chromium.org>