Refactor SubstitutionType for increased flexibility
This removes the SubstitutionType enum and the associated arrays of GN
and Ninja substitution names, replacing them with a more dynamic
collection of data.

This is part of a larger refactoring that will allow for additional
types of Tools (e.g. non-C/C++ langauge compilation) to be added more
simply and cleanly.

Change-Id: I2267a87b16d69790e584e729e426dab594784f71
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/4760
Commit-Queue: Julie Hockett <juliehockett@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>