Allow crate type to be set on all Rust targets
There are situations where the output type of the target does not match
the crate type in the compilation command. For example when the "--test"
flag is passed, an executable is produced, but the crate type is the
type of the specific crate, which is not necessarily 'bin'.

Change-Id: I7e2efc3ace9cbbc2ade838685967f7d26a2b21e6
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/5440
Commit-Queue: Julie Hockett <juliehockett@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>