Don't use a nested scope for BUILDCONFIG.
The change bac5bd0 ("Do not use the root scope for BUILDCONFIG.gn.")
introducted a nested scope into BUILDCONFIG.gn processing. But this
means that the build config file can't modify the default scope which is
its whole purpose.

This removes the nested scope and modifies the directory during
execution of the build config file which was the goal of the above
change.

Change-Id: I0790fae81d52554f9a80f0721aa9ae899a21c446
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/5880
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Brett Wilson <brettw@chromium.org>