Ignore SIGTRAP in child process on macOS
When running gn under Xcode (e.g. when debugging), the child process
is sent a SIGTRAP signal and immediately dies as it is not handled.
This makes gn consider that action, action_foreach and exec_script
have failed due to the signalled state of the child process.

To prevent this, ignore the SIGTRAP signal on macOS in the child
process.

Bug: none
Change-Id: Ic28312939d646db7d49c767d37566f3a8ed2c923
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/6662
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>