Fix Rust linking -lfoo.so not -lfoo due to TOC
Some toolchains generate a .so.TOC as well as a .so file when creating
shared libraries; they did not previously link correctly into downstream
Rust targets because the generated Ninja rules asked to link to -lfoo.so
rather than -lfoo.

The unit test code here is a bit verbose. It seemed desirable to test
with a toolchain that generates TOC files and one that does not.
However, that may be overkill.

Bug: crbug.com/gn/137
Change-Id: I30918957f8da187c5a1fc3e8f76d91fe5fda494f
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/7020
Reviewed-by: Adrian Taylor <adetaylor@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>