Add support for ':' in --gtest_filter to gn_unittests
For googletest, the ':' character can be used in either the posivite
or negative filter to separate a list of filters. Not having this in
gn_unittests is unexpected as it make it difficult to ignore multiple
tests suites (e.g. both ExecProcessTest and FormatTest fails when
building and running gn_unittests in Xcode).

Add the support for ':' by introducing another recursive function
PatternListMatchString and an early exit in PatternMatchesString.

Bug: none
Change-Id: I75709fb4c47981a1e9a9e4fb75295804ba87b38f
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/7061
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>