Add a sample project that allow to target iOS
This sample project allows building app bundles, framework bundles
and to build target on the host (e.g. to build a tool to preprocess
some file on the host before compiling them for iOS).

This sample does not support code-signature of the bundles (but it
will still embeds the necessary entitlements into the application
to allow debugging when targetting a simulator build).

This sample targets iOS 13.0+ SDK but can be adapted to target
older SDKs (only the demo application depends on a recent version
of the SDK).

Bug: 21
Change-Id: I80dff2c8eba989f84073bc3dd35326939b307080
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/6661
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>