Allow labels to go above the source root.
Supports paths in labels that go above the source root. These will then
be expanded to the absolute location of the source root on the local
system, and the relative directories will be expanded from there. This
patches the behavior for file path handling.

For example, if the current directory is "//" and the source root is
"/home/me/project", the strings "../..:foo" and "//../..:foo" will
resolve to "/home:foo".

Change-Id: I5e0a73105a24d91c9c6dc8905663421493caffe6
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/7480
Reviewed-by: Scott Graham <scottmg@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>