Make CONFIGURATION_BUILD_DIR represent location of binaries
Generation of CONFIGURATION_BUILD_DIR was hardcoded before. Because of
it, if bundle or binary builds to another directory it cause an issue when we try to
run target from XCode.
In this CL we get the real path to bundle and store it in CONFIGURATION_BUILD_DIR

Change-Id: I0347e5ad8b27e1ab287d66960751306f2f42faad
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/7640
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>