Remove flaky StringAtom test.
The test assumed that it was the only thing that initialized StringAtom
objects during testing, but this is no longer true now that StringAtom
are used all over the place, and allocated by other tests, which affect
the global StringAtom set.

Fix the issue by storing the input StringAtom pointers in place
before iterating over the set.

Change-Id: I65d56ad8933cf5984ef1b738a6c4d082c577acdd
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/7680
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>