Fix formatting of PBXContainerItemProxy
Ensure the formatting of PBXContainerItemProxy is identical to
the one used by Xcode to make comparison of generated project
with the upgraded version easier.

The object needs to be formatted on multiple line and the value
of remoteGlobalIDString should be just the object id, without
its comment (i.e. id() and not Reference()).

Bug: chromium/1071055
Change-Id: Id9e55777c67e90b03a52d7585b3f4c8a65afe48f
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/8083
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>