Set expected values by Xcode 11.4
If those values are not set, Xcode 11.4 consider the project as
requiring updating and it displays a warning to the user.

It can confuse developers experienced with having Xcode the source
of truth (i.e. experienced iOS developers but new to Chrome). As
the values are not used in anyway, set them so that Xcode does not
display a warning to developers.

Bug: chromium/1071055
Change-Id: Ia349daf6f6913e63990902ef87b92a6d6057c53b
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/8085
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>