Force the use of "Legacy Build System" in generated Xcode projects
Xcode 10.0+ supports two build systems named "Legacy Build System"
and "New Build System". Trying to build the project using the "New
Build System" results in cryptic errors.

Force the use of the "Legacy Build System" as it is known to work
well with the genenerated ninja project.

As the configuration is done at the workspace level, generate an
inner workspace for each project to hold the configuration (this
allow users to open the project instead of the workspace).

Bug: chromium:1091876, chromium:852522
Change-Id: If2ad1c47d75677f532e90e411437b29c885d32e8
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/8920
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>