Fix escaping in compile_commands.json
In particular this case is broken

  defines = ["INCLUDE=<header.h>"]

This is escaped as

  -DINCLUDE=\\u003Cheader.h>

which JSON decodes as

  -DINCLUDE=\u003Cheader.h>

which still has the JSON escape sequence \u003C in it. The current code
JSON escapes twice, but the command is only parsed once as JSON.

JSON escaping isn't shell escaping. Add the simplified shell escaping
described in the spec. This changed the output slightly, to quote spaces
instead of escaping them, but that's in line with the spec (including
the example).

VS Code is happy with this output, and not the previous output.

Bug: gn:181
Change-Id: I52f2470178066b072c2729181c91b947972b5bc4
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/9420
Reviewed-by: Brett Wilson <brettw@chromium.org>
Commit-Queue: Brett Wilson <brettw@chromium.org>