[rust-project] Add "core" and "alloc" as explicit deps for crates
rust-analyzer does not recognize any crates as beloning to sysroot
when analyzing rust-project.json projects. Because of this, it's
necessary to add "core" and "alloc" as explicit dependencies to
every crate so they can be accessed. There is no harm in including
them when they are not used, and this path is actually recommended
by the rust-analyzer team.

Change-Id: Iad2fe822b03ebc4eebd4bfb5f96b6273c3ecb6ea
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/9440
Commit-Queue: Petr Hosek <phosek@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>