clang: Add support for C++ modules in deps
This adds support for {{module_deps}}. This expansion is to be used by
the "cxx" tool. When the deps of a target include dependencies on
modules (defined as a target that includes a .modulemap file, which is
built to a .pcm), module_deps will be filled out with the required flags
to depend on those modules.

The targets that are depended upon must currently have a (handwritten)
.modulemap file in their sources.

The dependency semantics (as described in NinjaCBinaryTargetWriterTest)
are:
- .modulemap are built to .pcm by cxx_module
- .pcm aren't linked against
- The .cc of a target that uses modules depend on the .pcm ("implicit"
dependency in ninja terminology)
- The .cc sources of targets that depend on a module have implicit
dependencies on the pcm of the dependent modules target as well.
- A .a or executable does not depend on the .pcm.

Bug: fuchsia:27276
Change-Id: I84c26975b93db71e5309ad607fce900fe2705f90
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/9602
Commit-Queue: Scott Graham <scottmg@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>
Reviewed-by: Petr Hosek <phosek@google.com>