Identify current toolchain on error
Currently errors from GN identify the source location that failed, but
not the toolchain. This can be confusing, and adding a

  print(current_toolchain)

is often helpful when debugging these errors. This patch changes GN to
include the toolchain in the error message whenever there's an error
in a non-default toolchain.

This produces output like the following:

  ERROR at //subdir/BUILD.gn:3:1 (//:toolchain): Assertion failed.
  assert(only_true_in_default_toolchain)
  ^-----
  See //BUILD.gn:51:14: which caused the file to be included.
      deps = [ "subdir" ]
               ^-------

Bug: 184

Change-Id: I8463f04cf94242c4f7c518d4840ae8f586b4e4fa
Reviewed-on: https://gn-review.googlesource.com/c/gn/+/9640
Commit-Queue: Brett Wilson <brettw@chromium.org>
Reviewed-by: Brett Wilson <brettw@chromium.org>